Add ExternalLoginPlugin type Fix logic issue with the different get available plugin methods Add Front end login to display login links Beginning of implementation of Spotify's login process

+++ web/src/main/java/com/ftpix/nowplaying/models/Config.java
@@ -28,15 +28,19 @@ public class Config {
         Path backup = CONFIG_FOLDER.resolve(CONFIG_NAME + ".back");
         try {
             Gson gson = new GsonBuilder().setPrettyPrinting().create();
-            Files.move(configFile, backup, StandardCopyOption.REPLACE_EXISTING);
+            if (Files.exists(backup)) {
+                Files.move(configFile, backup, StandardCopyOption.REPLACE_EXISTING);
+            }
 
             String jsonStr = gson.toJson(this);
             System.out.println(jsonStr);
             Files.write(configFile, jsonStr.getBytes(), StandardOpenOption.WRITE, StandardOpenOption.CREATE);
-        }catch (IOException e){
-            Files.move(backup, configFile, StandardCopyOption.REPLACE_EXISTING);
-            throw  e;
-        }finally {
+        } catch (IOException e) {
+            if (Files.exists(backup)) {
+                Files.move(backup, configFile, StandardCopyOption.REPLACE_EXISTING);
+            }
+            throw e;
+        } finally {
             Files.deleteIfExists(backup);
         }
